:memo: Add kdoc documentation
appのkdocを追加
