+++ .idea/misc.xml
@@ -62,6 +62,15 @@
         <entry key="../../../../layout/compose-model-1647845081315.xml" value="0.17947635135135134" />
         <entry key="../../../../layout/compose-model-1647962926416.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1647962932018.xml" value="0.3888888888888889" />
+        <entry key="../../../../layout/compose-model-1648634454669.xml" value="0.17060810810810811" />
+        <entry key="../../../../layout/compose-model-1648794365366.xml" value="0.3888888888888889" />
+        <entry key="../../../../layout/compose-model-1648794632590.xml" value="0.19467905405405406" />
+        <entry key="../../../../layout/compose-model-1648794763549.xml" value="0.3888888888888889" />
+        <entry key="../../../../layout/compose-model-1648795626042.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1648797103200.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1648797107639.xml" value="0.16427364864864866" />
+        <entry key="../../../../layout/compose-model-1648797112546.xml" value="0.3888888888888889" />
+        <entry key="../../../../layout/compose-model-1648797141420.xml" value="0.3888888888888889" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.18541666666666667" />
         <entry key="app/src/main/res/drawable/baseline_public.xml" value="0.21354166666666666" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.1" />
