+++ app/src/main/java/com/searchSummarizer/app/SearchSummarizerActivity.kt
@@ -23,6 +23,11 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.context.GlobalContext
 import org.koin.core.context.startKoin
 
+/**
+ * SearchSummarizerAppを構成するsingle activity
+ *
+ * @constructor SearchSummarizerActivityを起動します。
+ */
 class SearchSummarizerActivity : ComponentActivity() {
 
     private val browserViewModel: BrowserViewModel by viewModel()
@@ -49,12 +54,20 @@ class SearchSummarizerActivity : ComponentActivity() {
         }
     }
 
+    /**
+     * App終了時、browser履歴を保存します。
+     *
+     * @param outState Bundle
+     */
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
         browserViewModel.saveBrowserHistory()
     }
 }
 
+/**
+ * SearchSummarizerAppのNavGraphを構成します。
+ */
 @Composable
 fun SearchSummarizerApp() {
     val navController = rememberNavController()
@@ -64,12 +77,20 @@ fun SearchSummarizerApp() {
     }
 }
 
+/**
+ * StartUp画面のNavGraph
+ *
+ * @param navController NavController
+ */
 private fun NavGraphBuilder.addStartUpGraph(navController: NavController) {
     composable(route = Screen.StartUp.route) {
         StartUpScreen(navController)
     }
 }
 
+/**
+ * Browser画面のNavGraph
+ */
 private fun NavGraphBuilder.addBrowserGraph() {
     composable(route = Screen.Browser.route) {
         BrowseScreen()
