+++ app/src/main/java/com/searchSummarizer/app/browser/BrowserWebViewClient.kt
@@ -8,9 +8,23 @@ import android.webkit.WebView
 import android.webkit.WebViewClient
 import androidx.annotation.RequiresApi
 
+/**
+ * Browser機能のcustom WebViewClient
+ *
+ * @property vm BrowserViewModel
+ * @constructor Create empty Browser web view client
+ */
 class BrowserWebViewClient(
     private val vm: BrowserViewModel
 ) : WebViewClient() {
+
+    /**
+     * browser履歴を更新します。
+     *
+     * @param view
+     * @param url
+     * @param favicon
+     */
     override fun onPageStarted(view: WebView?, url: String?, favicon: Bitmap?) {
         super.onPageStarted(view, url, favicon)
         requireNotNull(url)
@@ -22,6 +36,12 @@ class BrowserWebViewClient(
         vm.backEnabled = urlHistoryItem.size > 1
     }
 
+    /**
+     * page tileを更新します。
+     *
+     * @param view
+     * @param url
+     */
     override fun onPageFinished(view: WebView?, url: String?) {
         super.onPageFinished(view, url)
         val title = view?.title
@@ -29,6 +49,13 @@ class BrowserWebViewClient(
         vm.titles[vm.tabIndex] = title
     }
 
+    /**
+     * error pageを表示します。
+     *
+     * @param view
+     * @param request
+     * @param error
+     */
     @RequiresApi(Build.VERSION_CODES.M)
     override fun onReceivedError(
         view: WebView?,
