+++ app/src/main/java/com/searchSummarizer/data/enumType/HttpRoutes.kt
@@ -1,9 +1,26 @@
 package com.searchSummarizer.data.enumType
 
-sealed class HttpRoutes(val url: String) {
+/**
+ * host通信のhttp route
+ *
+ * @property url http routeのurl
+ * @constructor HttpRouteを作成します。
+ */
+sealed class HttpRoutes(private val url: String) {
 
+    /**
+     * host先のbase url
+     *
+     * @constructor BaseUrlを作成します。
+     */
     private object BaseUrl : Urls("https://search-summarizer.herokuapp.com")
 
-    data class SummarizedUrl(private val id: String) :
+    /**
+     * 共有されたSummarizedUrl
+     *
+     * @property id SummarizedUrlのid
+     * @constructor SharedSummarizedUrlを作成します。
+     */
+    data class SharedSummarizedUrl(private val id: String) :
         Urls("${BaseUrl.url}/summarized/urls?id=$id")
 }
