+++ app/src/main/java/com/searchSummarizer/data/model/BrowserData.kt
@@ -4,6 +4,14 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * browser履歴
+ *
+ * @property selectedTabIndex
+ * @property titles
+ * @property urls
+ * @constructor Create empty Browser history
+ */
 @Serializable
 data class BrowserHistory(
     @SerialName("selected_tab_index")
@@ -12,6 +20,16 @@ data class BrowserHistory(
     var urls: String = ""
 )
 
+/**
+ * まとめれたURLの情報
+ *
+ * @property id urlのid
+ * @property name url名
+ * @property author 作成者
+ * @property titles tabのtitle
+ * @property urls tabのurl
+ * @constructor SummarizedUrlを作成します。
+ */
 @Serializable
 data class SummarizedUrl(
     var id: String = "",
@@ -21,7 +39,14 @@ data class SummarizedUrl(
     var urls: List<String> = listOf()
 )
 
-
+/**
+ * browserのoption
+ *
+ * @property name option名
+ * @property icon icon
+ * @property onOptionSelected 選択されたときのevent
+ * @constructor BrowserOptionを作成します。
+ */
 data class BrowserOption(
     val name: String,
     val icon: ImageVector,
