+++ app/src/main/java/com/searchSummarizer/data/repo/context/ContextRepositoryImpl.kt
@@ -2,6 +2,12 @@ package com.searchSummarizer.data.repo.context
 
 import android.content.Context
 
+/**
+ * ContextRepositoryの実装
+ *
+ * @property context
+ * @constructor Create empty Context repository impl
+ */
 class ContextRepositoryImpl(
     private val context: Context
 ) : ContextRepository {
