+++ app/src/main/java/com/searchSummarizer/di/ViewModelModule.kt
@@ -5,6 +5,9 @@ import com.searchSummarizer.data.repo.browser.BrowserRepositoryImpl
 import com.searchSummarizer.data.repo.context.ContextRepositoryImpl
 import org.koin.dsl.module
 
+/**
+ * ViewModelのdi module
+ */
 val viewModelModule = module {
     single {
         BrowserViewModel(
